+++ lib/open890_web/components/rit_xit.ex
@@ -24,6 +24,13 @@ defmodule Open890Web.Components.RitXit do
                 <.cmd_button_2 cmd="XT1" classes="mini inverted secondary rit_xit">XIT</.cmd_button_2>
               <% end %>
             </span>
+            <span class="indicator">
+              <%= if @split_enabled do %>
+                <.cmd_button_2 cmd="TB0" classes="mini split-button enabled">SPLIT</.cmd_button_2>
+              <% else %>
+                <.cmd_button_2 cmd="TB1" classes="mini inverted secondary split-button">SPLIT</.cmd_button_2>
+              <% end %>
+            </span>
           </div>
 
           <div class="column two wide right aligned">